Add KEEP_ON_FAILURE flag to e2e kind tests (#832)
Using this flag allows debugging the failed test on the environment,
instead of destroying it.
The default behavior is still to clean up the cluster when failing.

Fixes: https://github.com/istio-ecosystem/sail-operator/issues/675

+++ tests/e2e/controlplane/control_plane_test.go
@@ -240,6 +240,10 @@ spec:
 					})
 
 					AfterAll(func(ctx SpecContext) {
+						if CurrentSpecReport().Failed() && keepOnFailure {
+							return
+						}
+
 						By("Deleting sample")
 						Expect(k.DeleteNamespace(sampleNamespace)).To(Succeed(), "sample namespace failed to be deleted")
 						Success("sample deleted")
@@ -281,6 +285,9 @@ spec:
 			if CurrentSpecReport().Failed() {
 				common.LogDebugInfo(common.ControlPlane, k)
 				debugInfoLogged = true
+				if keepOnFailure {
+					return
+				}
 			}
 
 			By("Cleaning up the Istio namespace")
@@ -294,9 +301,15 @@ spec:
 	})
 
 	AfterAll(func() {
-		if CurrentSpecReport().Failed() && !debugInfoLogged {
-			common.LogDebugInfo(common.ControlPlane, k)
-			debugInfoLogged = true
+		if CurrentSpecReport().Failed() {
+			if !debugInfoLogged {
+				common.LogDebugInfo(common.ControlPlane, k)
+				debugInfoLogged = true
+			}
+
+			if keepOnFailure {
+				return
+			}
 		}
 	})
 })
